MapTest Added
I've added a folder for the map generation testing. Main file to execute
is Tester.

The Floor class is kind of messy and I see a few ways to break it down
into smaller methods, but the code needed to work first.

This doesn't implement any gui as of yet, but I think this should at
least work as a base.