Updated Map (No GUI)
I added in the modified map classes, with a functional dungeon. I
commented out the section that runs it, since if it does run, then it
has to complete before the GUI element kicks in. It's messy, yes. Though
that's mainly since that code is for testing. Most of it won't be useful
in a GUI.