Player movement finished
The player now stays in the center of the screen, and the tile size was
increased so that only the area around the player shows.

I also decreased the max size of the maze, so the biggest it'll
currently get is 6x6. This is mostly since it's easier to get lost in
bigger sized mazes, especially when you can't see all of it.