Map is fully functional
The player can now loot from treasure chests, and run into random
battles.

The GUI was also fixed up to display what the buttons are for, and shows
which floor the player is on.

The Risk button gives the player a chance to gain equipment for free, or
encounter an enemy. I think it'd be better for it give a higher level
item, with a chance of fighting a higher level monster, but I'm not sure
how to switch the combat around without messing up the current level.

I'll try to get some graphics in as well.