Graphical Map Update
I added graphics to make things a little more impressive looking. All
the different difficulty of floors look different, so to show that
things are actually changing.

The player has a simple sprite that just faces the direction that you
last went, and the boss is some knight dude I found somewhere.